Fix http includes and move upgrade folders
