+++ src/daemons/MetaDaemon.cpp
@@ -19,9 +19,9 @@
 #include "kvstore/PartManager.h"
 #include "kvstore/NebulaStore.h"
 #include "meta/MetaServiceHandler.h"
-#include "meta/MetaHttpIngestHandler.h"
-#include "meta/MetaHttpDownloadHandler.h"
-#include "meta/MetaHttpReplaceHostHandler.h"
+#include "meta/http/MetaHttpIngestHandler.h"
+#include "meta/http/MetaHttpDownloadHandler.h"
+#include "meta/http/MetaHttpReplaceHostHandler.h"
 #include "meta/KVBasedClusterIdMan.h"
 #include "meta/ActiveHostsMan.h"
 #include "meta/processors/job/JobManager.h"
