+++ src/meta/CMakeLists.txt
@@ -92,13 +92,6 @@ nebula_add_library(
 )
 
 nebula_add_library(
-    meta_http_handler OBJECT
-    http/MetaHttpIngestHandler.cpp
-    http/MetaHttpDownloadHandler.cpp
-    http/MetaHttpReplaceHostHandler.cpp
-)
-
-nebula_add_library(
     meta_version_man_obj OBJECT
     MetaVersionMan.cpp
 )
@@ -108,5 +101,52 @@ add_dependencies(
     old_meta_thrift_obj
 )
 
+set(meta_test_deps
+    $<TARGET_OBJECTS:mock_obj>
+    $<TARGET_OBJECTS:meta_service_handler>
+    $<TARGET_OBJECTS:general_storage_service_handler>
+    $<TARGET_OBJECTS:storage_admin_service_handler>
+    $<TARGET_OBJECTS:graph_storage_service_handler>
+    $<TARGET_OBJECTS:storage_transaction_executor>
+    $<TARGET_OBJECTS:graph_storage_client_obj>
+    $<TARGET_OBJECTS:storage_client_base_obj>
+    $<TARGET_OBJECTS:internal_storage_client_obj>
+    $<TARGET_OBJECTS:storage_common_obj>
+    $<TARGET_OBJECTS:kvstore_obj>
+    $<TARGET_OBJECTS:raftex_obj>
+    $<TARGET_OBJECTS:wal_obj>
+    $<TARGET_OBJECTS:disk_man_obj>
+    $<TARGET_OBJECTS:keyutils_obj>
+    $<TARGET_OBJECTS:meta_client_obj>
+    $<TARGET_OBJECTS:file_based_cluster_id_man_obj>
+    $<TARGET_OBJECTS:ws_obj>
+    $<TARGET_OBJECTS:ws_common_obj>
+    $<TARGET_OBJECTS:http_client_obj>
+    $<TARGET_OBJECTS:raftex_thrift_obj>
+    $<TARGET_OBJECTS:storage_thrift_obj>
+    $<TARGET_OBJECTS:meta_thrift_obj>
+    $<TARGET_OBJECTS:common_thrift_obj>
+    $<TARGET_OBJECTS:thrift_obj>
+    $<TARGET_OBJECTS:process_obj>
+    $<TARGET_OBJECTS:meta_obj>
+    $<TARGET_OBJECTS:stats_obj>
+    $<TARGET_OBJECTS:fs_obj>
+    $<TARGET_OBJECTS:time_obj>
+    $<TARGET_OBJECTS:network_obj>
+    $<TARGET_OBJECTS:thread_obj>
+    $<TARGET_OBJECTS:conf_obj>
+    $<TARGET_OBJECTS:codec_obj>
+    $<TARGET_OBJECTS:datatypes_obj>
+    $<TARGET_OBJECTS:base_obj>
+    $<TARGET_OBJECTS:expression_obj>
+    $<TARGET_OBJECTS:function_manager_obj>
+    $<TARGET_OBJECTS:agg_function_manager_obj>
+    $<TARGET_OBJECTS:time_utils_obj>
+    $<TARGET_OBJECTS:encryption_obj>
+    $<TARGET_OBJECTS:ft_es_storage_adapter_obj>
+    $<TARGET_OBJECTS:version_obj>
+)
+
+nebula_add_subdirectory(http)
 nebula_add_subdirectory(test)
-nebula_add_subdirectory(upgradeData)
+nebula_add_subdirectory(upgrade)
