+++ src/meta/test/CMakeLists.txt
@@ -1,48 +1,7 @@
-set(meta_test_deps
-    $<TARGET_OBJECTS:mock_obj>
-    $<TARGET_OBJECTS:meta_service_handler>
-    $<TARGET_OBJECTS:general_storage_service_handler>
-    $<TARGET_OBJECTS:storage_admin_service_handler>
-    $<TARGET_OBJECTS:graph_storage_service_handler>
-    $<TARGET_OBJECTS:storage_transaction_executor>
-    $<TARGET_OBJECTS:graph_storage_client_obj>
-    $<TARGET_OBJECTS:storage_client_base_obj>
-    $<TARGET_OBJECTS:internal_storage_client_obj>
-    $<TARGET_OBJECTS:storage_common_obj>
-    $<TARGET_OBJECTS:kvstore_obj>
-    $<TARGET_OBJECTS:raftex_obj>
-    $<TARGET_OBJECTS:wal_obj>
-    $<TARGET_OBJECTS:disk_man_obj>
-    $<TARGET_OBJECTS:keyutils_obj>
-    $<TARGET_OBJECTS:meta_client_obj>
-    $<TARGET_OBJECTS:file_based_cluster_id_man_obj>
-    $<TARGET_OBJECTS:ws_obj>
-    $<TARGET_OBJECTS:ws_common_obj>
-    $<TARGET_OBJECTS:http_client_obj>
-    $<TARGET_OBJECTS:raftex_thrift_obj>
-    $<TARGET_OBJECTS:storage_thrift_obj>
-    $<TARGET_OBJECTS:meta_thrift_obj>
-    $<TARGET_OBJECTS:common_thrift_obj>
-    $<TARGET_OBJECTS:thrift_obj>
-    $<TARGET_OBJECTS:process_obj>
-    $<TARGET_OBJECTS:meta_obj>
-    $<TARGET_OBJECTS:stats_obj>
-    $<TARGET_OBJECTS:fs_obj>
-    $<TARGET_OBJECTS:time_obj>
-    $<TARGET_OBJECTS:network_obj>
-    $<TARGET_OBJECTS:thread_obj>
-    $<TARGET_OBJECTS:conf_obj>
-    $<TARGET_OBJECTS:codec_obj>
-    $<TARGET_OBJECTS:datatypes_obj>
-    $<TARGET_OBJECTS:base_obj>
-    $<TARGET_OBJECTS:expression_obj>
-    $<TARGET_OBJECTS:function_manager_obj>
-    $<TARGET_OBJECTS:agg_function_manager_obj>
-    $<TARGET_OBJECTS:time_utils_obj>
-    $<TARGET_OBJECTS:encryption_obj>
-    $<TARGET_OBJECTS:ft_es_storage_adapter_obj>
-    $<TARGET_OBJECTS:version_obj>
-)
+# Copyright (c) 2021 vesoft inc. All rights reserved.
+#
+# This source code is licensed under Apache 2.0 License,
+# attached with Common Clause Condition 1.0, found in the LICENSES directory.
 
 nebula_add_test(
     NAME
@@ -143,40 +102,6 @@ nebula_add_test(
         gtest
 )
 
-nebula_add_test(
-    NAME
-        meta_http_download_test
-    SOURCES
-        MetaHttpDownloadHandlerTest.cpp
-    OBJECTS
-        $<TARGET_OBJECTS:storage_http_handler>
-        $<TARGET_OBJECTS:meta_http_handler>
-        ${meta_test_deps}
-    LIBRARIES
-        ${ROCKSDB_LIBRARIES}
-        ${THRIFT_LIBRARIES}
-        ${PROXYGEN_LIBRARIES}
-        wangle
-        gtest
-)
-
-
-nebula_add_test(
-    NAME
-        meta_http_ingest_test
-    SOURCES
-        MetaHttpIngestHandlerTest.cpp
-    OBJECTS
-        $<TARGET_OBJECTS:storage_http_handler>
-        $<TARGET_OBJECTS:meta_http_handler>
-        ${meta_test_deps}
-    LIBRARIES
-        ${ROCKSDB_LIBRARIES}
-        ${THRIFT_LIBRARIES}
-        ${PROXYGEN_LIBRARIES}
-        wangle
-        gtest
-)
 
 nebula_add_test(
     NAME
@@ -262,23 +187,6 @@ nebula_add_test(
         wangle
         gtest
 )
-
-#nebula_add_test(
-#    NAME
-#        meta_http_replace_test
-#    SOURCES
-#        MetaHttpReplaceHandlerTest.cpp
-#    OBJECTS
-#        $<TARGET_OBJECTS:storage_http_handler>
-#        $<TARGET_OBJECTS:meta_http_handler>
-#        ${meta_test_deps}
-#    LIBRARIES
-#        ${ROCKSDB_LIBRARIES}
-#        ${THRIFT_LIBRARIES}
-#        ${PROXYGEN_LIBRARIES}
-#        wangle
-#        gtest
-#)
 
 nebula_add_test(
     NAME
