+++ src/meta/upgrade/MetaDataUpgrade.cpp
@@ -18,8 +18,8 @@
 #include "kvstore/Common.h"
 #include "meta/MetaServiceUtils.h"
 #include "meta/ActiveHostsMan.h"
-#include "meta/upgradeData/MetaDataUpgrade.h"
-#include "meta/upgradeData/oldThrift/MetaServiceUtilsV1.h"
+#include "meta/upgrade/MetaDataUpgrade.h"
+#include "meta/upgrade/oldThrift/MetaServiceUtilsV1.h"
 
 DECLARE_bool(null_type);
 DECLARE_uint32(string_index_limit);
